requirements.txt
# Test Runner
pytest == 8.2.0
pytest-asyncio == 0.23.6
pytest-playwright == 0.4.4
pytest-html == 4.1.1
pytest-xdist == 3.5.0

# UI Automation
playwright == 1.44.0

# API Automation
httpx == 0.27.0
requests == 2.31.0

# Reporting
pytest-html == 4.1.1
allure-pytest == 2.13.5

# Utilities
python-dotenv == 1.0.1
pyyaml == 6.0.1

# Logging
loguru == 0.7.2

# Assertions
assertpy == 1.1

# Data Generation
faker == 22.0.0